requirements.txt
boto==2.49.0
boto3==1.9.71
botocore==1.12.71
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.4
django-crispy-forms==1.7.2
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
idna==2.8
jmespath==0.9.3
Pillow==5.3.0
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
stripe==2.19.0
urllib3==1.24.1
